+++ src/main/java/com/courtcanva/ccfranchise/models/CcEmployee.java
@@ -1,4 +1,4 @@
-package com.courtcanva.ccfranchise.model;
+package com.courtcanva.ccfranchise.models;
 
 import com.courtcanva.ccfranchise.constants.EmployeeStatus;
 import com.courtcanva.ccfranchise.constants.Role;
