+++ src/main/java/com/courtcanva/ccfranchise/repositories/StaffRepository.java
@@ -1,6 +1,6 @@
 package com.courtcanva.ccfranchise.repositories;
 
-import com.courtcanva.ccfranchise.model.Staff;
+import com.courtcanva.ccfranchise.models.Staff;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StaffRepository extends JpaRepository<Staff, Long> {
