+++ src/test/java/com/courtcanva/ccfranchise/services/StaffServiceTest.java
@@ -3,8 +3,8 @@ package com.courtcanva.ccfranchise.services;
 import com.courtcanva.ccfranchise.constants.State;
 import com.courtcanva.ccfranchise.dtos.StaffInfoDto;
 import com.courtcanva.ccfranchise.mappers.StaffMapper;
-import com.courtcanva.ccfranchise.model.Franchisee;
-import com.courtcanva.ccfranchise.model.Staff;
+import com.courtcanva.ccfranchise.models.Franchisee;
+import com.courtcanva.ccfranchise.models.Staff;
 import com.courtcanva.ccfranchise.repositories.StaffRepository;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
